use v-dialog inside the sub component
